+++ CHANGELOG.md
@@ -389,3 +389,4 @@
 * Define experimental `reconcile` method in `State` module
 * Remove `Maybe` from `Prefixed` type synonymous
 * Define `discover` for `Morphable Find` instances with composite keys
+* Change `subset` method of `Set` interface: returns `Convergence Boolean`